+++ core/templates/dev/head/pages/exploration_player/ExplorationPlayer.ts
@@ -61,7 +61,7 @@ require('components/background/BackgroundBannerDirective.ts');
 require('pages/exploration_player/ConversationSkinDirective.ts');
 require('pages/exploration_player/ExplorationFooterDirective.ts');
 require('pages/exploration_player/LearnerLocalNavDirective.ts');
-require('pages/exploration_player/LearnerViewInfo.ts');
+require('pages/exploration_player/LearnerViewInfoDirective.ts');
 
 require('domain/exploration/ReadOnlyExplorationBackendApiService.ts');
 require('services/ContextService.ts');
