+++ core/templates/dev/head/pages/exploration_player/exploration_player.html
@@ -29,16 +29,7 @@
 {% endblock header_js %}
 
 {% block navbar_breadcrumb %}
-  <ul class="nav navbar-nav oppia-navbar-breadcrumb" ng-controller="LearnerViewInfo">
-    <li>
-      <span class="oppia-navbar-breadcrumb-separator"></span>
-      <h1 class="oppia-exploration-h1"><span class="protractor-test-exploration-header oppia-exploration-header" itemprop="description">{{exploration_title}}</span></h1>
-    </li>
-    <li ng-click="showInformationCard()" uib-tooltip="<['I18N_PLAYER_INFO_TOOLTIP' | translate]>" tooltip-placement="bottom" style="cursor: pointer;" class="oppia-exploration-info-icon protractor-test-exploration-info-icon" tabindex="0">
-      <i class="material-icons oppia-navbar-breadcrumb-icon" style="font-size: 20px;">&#xE88E;</i>
-      <span class="oppia-icon-accessibility-label"><['I18N_PLAYER_INFO_TOOLTIP' | translate]></span>
-    </li>
-  </ul>
+  <learner-view-info></learner-view-info>
 {% endblock navbar_breadcrumb %}
 
 {% block local_top_nav_options %}
