+++ core/templates/dev/head/pages/exploration_player/learner_view_info_directive.html
@@ -1,0 +1,10 @@
+<ul class="nav navbar-nav oppia-navbar-breadcrumb">
+  <li>
+    <span class="oppia-navbar-breadcrumb-separator"></span>
+    <h1 class="oppia-exploration-h1"><span class="protractor-test-exploration-header oppia-exploration-header" itemprop="description">{{exploration_title}}</span></h1>
+  </li>
+  <li ng-click="$ctrl.showInformationCard()" uib-tooltip="<['I18N_PLAYER_INFO_TOOLTIP' | translate]>" tooltip-placement="bottom" style="cursor: pointer;" class="oppia-exploration-info-icon protractor-test-exploration-info-icon" tabindex="0">
+    <i class="material-icons oppia-navbar-breadcrumb-icon" style="font-size: 20px;">&#xE88E;</i>
+    <span class="oppia-icon-accessibility-label"><['I18N_PLAYER_INFO_TOOLTIP' | translate]></span>
+  </li>
+</ul>
