+++ core/templates/dev/head/pages/question_player/QuestionPlayerDirective.ts
@@ -63,7 +63,7 @@ require('components/background/BackgroundBannerDirective.ts');
 require('pages/exploration_player/ConversationSkinDirective.ts');
 require('pages/exploration_player/ExplorationFooterDirective.ts');
 require('pages/exploration_player/LearnerLocalNavDirective.ts');
-require('pages/exploration_player/LearnerViewInfo.ts');
+require('pages/exploration_player/LearnerViewInfoDirective.ts');
 
 oppia.directive('questionPlayer', [
   '$http', 'UrlInterpolationService',
